Import changes for goma client
  - c47c6114f13870c323f7ecfee63d88ecd86b59c8 cpp_parser: handle integer bigger than 32-bits
  - 0af37164491ccddef060ea731e938fd0d1857037 Roll boringssl from ed5f4e82e69d to f3e5941515dd (8 revis...
  - fd38cf41f7bf90f5091c517f964628170fb73eb2 Roll clang from 80c14f040262 to c8d8b3ecbd25 (1 revision)
  - 3f266a3c3fc3fee175f44eae220649d1d58dbf7c Roll clang from 83b09bd60932 to 80c14f040262 (1 revision)
  - 52040c319c7ee02d19e1e83a5cd064bf3cee99c9 Update build/config/mac_sdk.gni
  - e8a12dc65c3777dd82638a000f695a435b7b69a9 Roll clang from a442c61f6ee3 to 83b09bd60932 (1 revision)
  - 7af362838d7b60f546f586fd96b3d808a99d7403 Roll clang from 620e6a0475a2 to a442c61f6ee3 (1 revision)
  - c17a40805f7c02150b59fe480578cac248714e1d Roll clang from e74330fba0ba to 620e6a0475a2 (1 revision)
  - fef365b17354525f66aa5fdd106493cccce5bc71 generate_feature_list.go parses Diagnostic.td
  - 421421af7e231d091ede4a5669083e09573987c1 Update go to 1.17.1
  - b83c47d52ab35ca793211d6d58b5afee5b6c2625 Roll clang from 664e4259b150 to e74330fba0ba (1 revision)

GitOrigin-RevId: c47c6114f13870c323f7ecfee63d88ecd86b59c8

+++ client/cxx/include_processor/cpp_parser.cc
@@ -450,7 +450,7 @@ void CppParser::ProcessIfndef(const CppDirectiveIfndef& d) {
 
 void CppParser::ProcessIf(const CppDirectiveIf& d) {
   GOMA_COUNTERZ("if");
-  int v = EvalCondition(d.tokens());
+  int64_t v = EvalCondition(d.tokens());
   VLOG(2) << DebugStringPrefix() << " #IF " << v;
   conditions_.push_back(Condition(v != 0));
 }
@@ -495,7 +495,7 @@ void CppParser::ProcessElif(const CppDirectiveElif& d) {
     return;
   }
 
-  int v = EvalCondition(d.tokens());
+  int64_t v = EvalCondition(d.tokens());
   VLOG(2) << DebugStringPrefix() << " #ELIF " << v;
   conditions_.back().cond = (v != 0);
   conditions_.back().taken |= (v != 0);
@@ -654,7 +654,7 @@ void CppParser::ProcessIncludeInternal(const CppDirectiveIncludeBase& d) {
   Error("#include expects \"filename\" or <filename>");
 }
 
-int CppParser::EvalCondition(const ArrayTokenList& orig_tokens) {
+int64_t CppParser::EvalCondition(const ArrayTokenList& orig_tokens) {
   // TODO: Add DCHECK here orig_tokens does not contain spaces.
   ArrayTokenList tokens;
   tokens.reserve(orig_tokens.size());
